+++ src/App.tsx
@@ -12,9 +12,9 @@ import LostAndFound from './pages/LostAndFound'
 import AdoptionForm from './pages/AdoptionForm'
 import DonationForm from './pages/DonationForm'
 import VolunteerForm from './pages/VolunteerForm'
-import LostReport from './pages/LostReport'
-import FoundReport from './pages/FoundReport'
-import AbusedReport from './pages/AbusedReport'
+import LostReport from './pages/ReportForm/LostReport'
+import FoundReport from './pages/ReportForm/FoundReport'
+import AbusedReport from './pages/ReportForm/AbusedReport'
 import ReportModal from './components/ReportModal'
 import LoginModal from './components/LoginModal'
 import SignUpModal from './components/SignUpModal'
